Created division button controller and added it to the view Transform results type int to float

+++ src/main/java/com/costardstudio/patterns/features/observer/controller/MultiplicationButtonController.java
@@ -16,11 +16,11 @@ public class MultiplicationButtonController extends CalculatorButtonController {
     }
 
     @Override
-    public int computeOperation(List<String> inputs) {
-        int multiplicationResult = Integer.parseInt(inputs.get(0));
+    public float computeOperation(List<String> inputs) {
+        float multiplicationResult = Float.parseFloat(inputs.get(0));
         inputs.remove(0);
         for (String input : inputs) {
-            multiplicationResult = multiplicationResult * Integer.parseInt(input);
+            multiplicationResult = multiplicationResult * Float.parseFloat(input);
         }
         return multiplicationResult;
     }
